requirements.txt
minecraft-launcher-lib
customtkinter
pywinstyles
pillow